pyproject.toml
[tool.poetry]
name = "choi"
version = "0.1.0"
description = ""
authors = ["최재현 <zip020314@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
selenium = "^4.33.0"
djangorestframework = "^3.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
